+++ Gruntfile.js
@@ -1,18 +1,20 @@
 /*globals module */
 module.exports = function (grunt) {
 
+	var banner = '/*! <%= pkg.name %> - v<%= pkg.version %> - ' +
+		'<%= grunt.template.today("yyyy-mm-dd") %> */';
+
 	// Project configuration.
 	grunt.initConfig({
-		pkg: '<json:package.json>',
-		meta: {
-			banner: '/*! <%= pkg.name %> - v<%= pkg.version %> - ' +
-			'<%= grunt.template.today("yyyy-mm-dd") %> */'
-		},
+		pkg: require('./package'),
 
 		concat: {
+			options: {
+				banner: banner
+			},
 			build: {
 				src: ['partials/*.scss'],
-				dest: 'mixins.scss'
+				dest: 'mixins.scss',
 			}
 		},
 
@@ -20,7 +22,8 @@ module.exports = function (grunt) {
 			css: {
 				options: {
 					unixNewlines: true,
-					style: 'expanded'
+					style: 'expanded',
+					banner: banner
 				},
 				files: {
 					'tests/css/main.css': 'tests/scss/main.scss'